Change int to long

+++ src/Main.java
@@ -1,8 +1,8 @@
 public class Main {
     public static void main(String[] args) {
-        SalesManager salesManager = new SalesManager(new int[] {100, 200, 50, 10});
+        SalesManager salesManager = new SalesManager(new long[] {100, 200, 50, 10});
 
-        int max = salesManager.max();
+        long max = salesManager.max();
         System.out.println("Максимальный размер продаж: " + max);
     }
 }
